Make all provided messages push in single grpc call
